Ensure the first demo is fully successful
`intro-demo` from chapter1 regularly does not reach 100% accuracy in
my testing. The text in the notebook seems to imply this is because of
the size of the input. However, given that the input contains the
entirety of the problem space, I believed we should be able to
reliably provide a demo that produces 100% accuracy.

I tried a number of changes, but settled on increasing the size of the
hidden layer as having the most significant impact on the result. By
going from a size of 4 to 32, I now get 100% accuracy reliably with
less than 200 epochs, and usually far fewer than that. The code now
does it 10 epochs at a time and stops when it's at 100%.

I've also dropped the text that tried to explain why the accuracy may
be low, as we should now always have an accurate result.

This was a helpful post where someone else explained the steps they
went through for the same example and which variables produced the
most impact for them.

    https://blog.thoughtram.io/machine-learning/2016/11/02/understanding-XOR-with-keras-and-tensorlow.html

Many thanks to Chris Sibbit who helped me understand many details
about neutral network structure, loss functions, optimization
algorithms, and debugging techniques based on his experience.

Co-authored-by: Chris Sibbitt <csibbitt@redhat.com>
Signed-off-by: Russell Bryant <rbryant@redhat.com>